+++ src/components/common/Layout.tsx
@@ -1,0 +1,20 @@
+import { Header } from '../Header';
+
+import { PropsWithChildren } from 'react';
+import { styled } from 'styled-components';
+
+export const Layout = ({ children }: PropsWithChildren) => (
+  <Style.Layout>
+    <Header />
+    {children}
+  </Style.Layout>
+);
+
+const Style = {
+  Layout: styled.div`
+    display: flex;
+    flex-direction: column;
+
+    align-items: center;
+  `,
+};
